+++ react/components/organisms/listing/listing.js
@@ -25,7 +25,10 @@ export class Listing extends Component {
             emptyItemsText: PropTypes.string,
             search: PropTypes.bool,
             loading: PropTypes.bool,
+            refreshing: PropTypes.bool,
             flatListProps: PropTypes.object,
+            onEndReachedThreshold: PropTypes.number,
+            onFilter: PropTypes.func,
             style: ViewPropTypes.style,
             scrollViewStyle: ViewPropTypes.style,
             scrollViewContainerStyle: ViewPropTypes.style
@@ -41,11 +44,9 @@ export class Listing extends Component {
             emptyItemsText: "No items",
             search: true,
             loading: false,
+            refreshing: false,
             flatListProps: {},
-            onSearch: async () => {},
             onFilter: async () => {},
-            onRefresh: async () => {},
-            onEndReached: async () => {},
             style: {},
             scrollViewStyle: {},
             scrollViewContainerStyle: {}
@@ -102,6 +103,7 @@ export class Listing extends Component {
     };
 
     onFilter = async value => {
+        await this.props.onFilter(value);
         this.setState({ filters: value }, async () => await this.refresh());
     };
 
@@ -250,17 +252,17 @@ export class Listing extends Component {
                     key={"items"}
                     style={styles.flatList}
                     data={this.state.items}
-                    refreshing={this.state.refreshing}
+                    refreshing={this.props.refreshing && this.state.refreshing}
                     onRefresh={this.onRefresh}
                     onEndReached={this.onEndReached}
-                    onEndReachedThreshold={0.6}
+                    onEndReachedThreshold={this.props.onEndReachedThreshold}
                     renderItem={({ item, index }) => this.props.renderItem(item, index)}
                     keyExtractor={item => String(item.id)}
                     ListEmptyComponent={this._renderEmptyList()}
                     ListFooterComponent={<View style={styles.flatListBottom} />}
                     {...this.props.flatListProps}
                 />
-                {this.state.loading && (
+                {this.props.loading && this.state.loading && (
                     <ActivityIndicator
                         style={styles.loadingIndicator}
                         size="large"